Revert "Vehicle Parts refactored"
This reverts commit 787a70fee67685041e6ad29d0c5e969e86004b50.
